Fix & Refactor : git conflict 오류 수정 및  live code review feedback 반영 함수형 state 만들기, main에서만 쓰이는 component main 안으로 위치 변경, 불필요한 변수 삭제 등
