+++ public/data/Hyojeong/feedData.json
@@ -5,7 +5,7 @@
     "userName": "Happy_day",
     "content": "What a glory day🌼",
     "thumbnail": "https://images.unsplash.com/photo-1489710437720-ebb67ec84dd2?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1470&q=80",
-    "comments": [
+    "commentList": [
       {
         "id": 1,
         "userName": "wecode",
@@ -33,7 +33,7 @@
     "content": "Hi there.",
     "thumbnail": "https://images.unsplash.com/photo-1532635241-17e820acc59f?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1415&q=80",
 
-    "comments": [
+    "commentList": [
       {
         "id": 1,
         "userName": "sally_world",
@@ -60,7 +60,7 @@
     "userName": "jayPark",
     "content": "the SKY",
     "thumbnail": "https://images.unsplash.com/photo-1618672853783-a85ce33df16c?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1475&q=80",
-    "comments": [
+    "commentList": [
       {
         "id": 1,
         "userName": "we___together",
@@ -88,7 +88,7 @@
     "userName": "Sally",
     "content": "홀로 작업중 ✒️",
     "thumbnail": "https://images.unsplash.com/photo-1553171403-619f44bdb0aa?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=2188&q=80",
-    "comments": [
+    "commentList": [
       {
         "id": 1,
         "userName": "we_healing",
@@ -116,7 +116,7 @@
     "content": "포로리야",
     "thumbnail": "https://scontent-ssn1-1.xx.fbcdn.net/v/t1.18169-9/14502788_309919656067277_8807121396242408041_n.jpg?_nc_cat=101&ccb=1-5&_nc_sid=9267fe&_nc_ohc=euMvF-3OBWAAX8hcVLR&_nc_ht=scontent-ssn1-1.xx&oh=00_AT_lHnLv89boi30F5Nbgn9axgP6YDrxq7sMmH3Gphnq_IA&oe=625EB063",
 
-    "comments": [
+    "commentList": [
       {
         "id": 1,
         "userName": "locingMy_12",
